preference_webapp/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
gpytorch>=1.11.0
pytorch-lightning>=2.0.0
rasterio>=1.3.0
matplotlib>=3.7.0
plotly>=5.17.0
scikit-learn>=1.3.0
pandas>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

# Deployment dependencies
gunicorn==21.2.0
